pyproject.toml
[project]
name = "neumann"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "mypy>=1.17.1",
    "openai>=1.99.9",
    "pdf2image>=1.17.0",
    "pillow>=10.0.0",
    "pymupdf>=1.26.3",
    "requests>=2.30.0",
]
